fix: remove separate mweb stream file
